requirements.txt
httpx==0.27.0
pydantic==2.6.3
fastapi==0.110.0
uvicorn==0.27.1
google-genai
python-dotenv